Refactor: seperate types from components
