+++ src/components/Auth/Input/InputEmail.tsx
@@ -1,27 +1,10 @@
-import { UseFormRegister, UseFormResetField } from "react-hook-form";
-
 import styles from "./Input.module.scss";
 
 import RemoveBtn from "@/assets/icons/removeBtn.svg?react";
 
-interface LoginForm {
-  email: string;
-  password: string;
-}
+import { LoginInput } from "@/types/auth";
 
-interface LoginDirtyFields {
-  email?: boolean;
-  password?: boolean;
-}
-
-interface Props {
-  label: string;
-  dirtyFields: LoginDirtyFields;
-  register: UseFormRegister<LoginForm>;
-  resetField: UseFormResetField<LoginForm>;
-}
-
-function InputEmail({ label, register, dirtyFields, resetField }: Props) {
+function InputEmail({ label, register, dirtyFields, resetField }: LoginInput) {
   const resetEmail = () => {
     resetField("email");
   };
